Add sort-operator-numeric
It's too complicated to support true numeric sort char-wise, but very
simple and useful to support it line-wise. So do a basic version for
char-wise that converts everything to a number.

Change the sort to be stable since characters after a number make sorted
objects not uniquely identified by the sort key (the number).

Using <Plug>(sort-operator-numeric)ib, we get:

    (bb, aa, 45, cc, 54, 5, 39, 40aa, -10)
sorted to:
    (-10, bb, aa, cc, 5, 39, 40aa, 45, 54)
sorted to:
    (-10, bb, aa, cc, 5, 39, 40aa, 45, 54)

`:sort n` would give:
    bb, aa, cc, -10, 5, 39, 40aa, 45, 54,

The sort is stable, and sorts numbers but not strings. It positions
strings where 0 fits in the list of numbers.

In contrast, using <Plug>(sort-operator)ib, we get:

    (bb, aa, 45, cc, 54, 5, 39, 40aa, -10)
sorted to:
    (-10, 39, 40aa, 45, 5, 54, aa, bb, cc)
sorted to:
    (-10, 39, 40aa, 45, 5, 54, aa, bb, cc)